# api_documentation/Core.swagger_collection.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Deep Lynx",
...
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/containers/{container-id}": {
      "put": {
        "description": "Updates the container. This will fail if a container already exists with the proposed updated name.",
        "summary": "Update",
        "tags": [
          "Containers"
        ],
        "operationId": "Update",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Fetch container by ID.",
        "summary": "Fetch by ID",
        "tags": [
          "Containers"
        ],
        "operationId": "FetchbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives a Container. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive",
        "tags": [
          "Containers"
        ],
        "operationId": "Archive",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/containers/{container-id}/metatypes/{type-id}/keys": {
      "post": {
        "description": "Creates a new key for a type. Keys consist of a unique key name (unique to the type not as a whole), key type, default values, and allowed values. Of those, only the first two are required.\n\nThe `dataType` field accepts only one of the following values: number, string, date, boolean, enumeration. \n\nThe fields `defaultValue` and `options` will only accept an array of the following types: string, boolean, number, float.\n\nPass in an array for bulk creation.\n\nCurrently the validation and cardinality functionality of keys are NOT checked at data insertion.",
        "summary": "New Type Key",
        "tags": [
          "Metatype Keys"
        ],
        "operationId": "NewTypeKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewTypeKeyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Lists all currently valid and available keys for the type. The response includes the key name, its types, allowed values, and default value if present. MUST include the limit/offset query parameters as of 3/26/2020",
        "summary": "List Type's Keys",
        "tags": [
          "Metatype Keys"
        ],
        "operationId": "ListType'sKeys",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/metatypes/{type-id}/keys/{key-id}": {
      "get": {
        "description": "Lists all currently valid and available keys for the type. The response includes the key name, its types, allowed values, and default value if present.",
        "summary": "Get Type Key",
        "tags": [
          "Metatype Keys"
        ],
        "operationId": "GetTypeKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "key-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "Updates a single key for a type.",
        "summary": "Update Type Key",
        "tags": [
          "Metatype Keys"
        ],
        "operationId": "UpdateTypeKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "key-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateTypeKeyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archiving the type key prevents any new types from implementing the key. It *does not remove key/value pairs on existing types*. We highly recommend you archive type keys instead of deleting them so that previous data is not affected.",
        "summary": "Archive Type Key",
        "tags": [
          "Metatype Keys"
        ],
        "operationId": "ArchiveTypeKey",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "type-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "key-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateRequest": {
      "title": "UpdateRequest",
      "example": {
        "description": "Final Test Description"
      },
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        }
      },
      "required": [
        "description"
      ]
    },
...
    "NewTypeKeyRequest": {
      "title": "NewTypeKeyRequest",
      "example": {
        "name": "File Name",
        "required": true,
        "property_name": "fileName",
        "description": "File Upload Name",
        "data_type": "string",
        "cardinality": 1,
        "validation": {
          "regex": "",
          "min": 0,
          "max": 0
        },
        "unique": true,
        "options": [
          ""
        ],
        "defaultValue": "can be any type"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "required": {
          "type": "boolean"
        },
        "property_name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "data_type": {
          "type": "string"
        },
        "cardinality": {
          "type": "integer",
          "format": "int32"
        },
        "validation": {
          "$ref": "#/definitions/Validation"
        },
        "unique": {
          "type": "boolean"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "defaultValue": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "required",
        "property_name",
        "description",
        "data_type",
        "cardinality",
        "validation",
        "unique",
        "options",
        "defaultValue"
      ]
    },
    "Validation": {
      "title": "Validation",
      "example": {
        "regex": "",
        "min": 0,
        "max": 0
      },
      "type": "object",
      "properties": {
        "regex": {
          "type": "string"
        },
        "min": {
          "type": "integer",
          "format": "int32"
        },
        "max": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "regex",
        "min",
        "max"
      ]
    },
    "UpdateTypeKeyRequest": {
      "title": "UpdateTypeKeyRequest",
      "example": {
        "name": "Updated String",
        "property_name": "string",
        "description": "sting",
        "data_type": "string",
        "default_value": "optional - data structure here is dependent on dataType value"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "property_name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "data_type": {
          "type": "string"
        },
        "default_value": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "property_name",
        "description",
        "data_type",
        "default_value"
      ]
    },
...
  },
...
}